Rename `pot_aes` module to simply `aes`
